pyproject.toml
[build-system]
requires = ["hatchling>=1.21.1"]
build-backend = "hatchling.build"

[project]
name = "codegeni"
version = "0.1.0"
description = "CodeGeni - AI-assisted code review engine and CLI"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "SolutionMind", email = "" }
]
keywords = ["code-review", "lint", "security", "ast", "vscode"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "click>=8.1.7",
  "pydantic>=2.7",
  "PyYAML>=6.0.1",
]

[project.urls]
Homepage = "https://example.com/codegeni"
Repository = "https://example.com/codegeni/repo"

[project.optional-dependencies]
analyzers = [
  "flake8>=6.1.0",
  "pylint>=3.2.0",
  "bandit>=1.7.8",
  "radon>=6.0.1",
]
server = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
  "black>=24.0",
  "mypy>=1.10",
  "types-PyYAML>=6.0.12",
]

[project.scripts]
codegeni = "codegeni.cli:main"

[tool.codegeni]
# Default configuration values; can be overridden in .codegeni.yaml or pyproject.toml [tool.codegeni]
enable = true
analyzers = ["flake8", "pylint", "bandit", "radon", "python-ast"]
severity_threshold = "info"
exclude = [".git", ".venv", "node_modules", "dist", "build"]
python = { max_function_lines = 60, max_nesting_depth = 3 }


[tool.hatch.build.targets.wheel]
packages = ["src/codegeni"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
extend-exclude = ["node_modules", "dist", "build", ".venv"]

[tool.black]
line-length = 100


